requirements.txt
fastapi
uvicorn[standard]
google-cloud-texttospeech
python-multipart   # only if you accept form uploads
pydantic
